==> 1_Quality_control.Rmd <==
---
title: "GSEA analysis"
author: "Kelly"
date: "12/05/2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```
# Loading packages

```{r packages}
library(readr)
library(ggplot2)
library(dplyr)
library(ggrepel)
library(limma)
library(tibble)
library(magrittr)
library(dplyr)
library(reshape2)
```

# Data import

```{r}
setwd("~/Documents/Kelly_2021/Research/Tian_Camel_PhD/Kelly_iTRAQ_Data_analysis_11242021")
QC_report <- read.csv("iTRAQ_QC.csv")
QC_report <- QC_report[,1:6]
```

```{r}
melt(QC_report[2])%>%
  rbind(melt(QC_report[3]))%>%
  rbind(melt(QC_report[4]))%>%
  rbind(melt(QC_report[5]))%>%
  mutate(File = rep(c(1:20), 4))%>%
  set_colnames(c("variable", "number", "File"))%>%
  ggplot(aes(x=File, y=number, fill=variable)) +
  geom_bar(stat="identity") + 
  theme_bw()
```
